chore: ui tweaks
