+++ src/components/Icons.tsx
@@ -859,7 +859,7 @@ export const Icons = {
           <path
             fillRule="evenodd"
             clipRule="evenodd"
-            d="M15.3303 15.748C15.3303 16.4442 15.8873 17.0012 16.5835 17.0012C17.2797 17.0012 17.8367 16.4442 17.8367 15.748C17.8367 15.0518 17.2797 14.4949 16.5835 14.4949C15.8873 14.4949 15.3303 15.0518 15.3303 15.748Z"
+            d="M15.3303 15.748C15.3303 16.4442 15.8873 17.0012 16.5835 17.0012C17.2797 17.0012 17.8367 16.4442 17.8367 15.748C17.8367 15.0svg 17.2797 14.4949 16.5835 14.4949C15.8873 14.4949 15.3303 15.0518 15.3303 15.748Z"
             fill="white"
           />
         </g>
