+++ src/components/defi-card.tsx
@@ -1,8 +1,9 @@
-import React from "react";
 import Link from "next/link";
+import React from "react";
+
 import { Button } from "@/components/ui/button";
-import { cn } from "@/lib/utils";
 import { Skeleton } from "@/components/ui/skeleton";
+import { cn } from "@/lib/utils";
 
 export interface TokenDisplay {
   icon: React.ReactNode;
